[#94][Refactor] 쇼츠 좋아요 api 분리
