+++ src/main/java/com/adregamdi/like/dto/response/CreateShortsLikeResponse.java
@@ -1,0 +1,6 @@
+package com.adregamdi.like.dto.response;
+
+public record CreateShortsLikeResponse(
+        int likeCnt
+) {
+}
